% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cocktail.R
\name{theme_cocktail}
\alias{theme_cocktail}
\title{A custom theme for cocktail visualizations}
\usage{
theme_cocktail()
}
\value{
A ggplot2 theme.
}
\description{
Creates a clean, polished theme for plots made with the
cocktailtools package.
}
